+++ src/main/java/game/race/support/CarRacingManager.java
@@ -3,23 +3,22 @@ package game.race.support;
 import java.util.List;
 
 import game.race.Car;
-import game.race.model.InputView;
-import game.race.model.ResultView;
+import game.race.dto.RaceDto;
+import game.race.view.ResultView;
 
 public class CarRacingManager {
 
-    public static void start(InputView inputView) {
+    public static void start(RaceDto raceDto) {
         ResultView resultView = new ResultView();
-        List<Car> cars = Car.getCars(inputView.getVehicleCnt());
+        List<Car> cars = Car.getCars(raceDto.getVehicleCnt());
 
         resultView.showPrompt();
-        for (int trial = 0; trial < inputView.getTryCnt(); trial++) {
+        for (int trial = 0; trial < raceDto.getTryCnt(); trial++) {
             for (Car car : cars) {
                 car.move();
             }
 
             resultView.showCars(cars);
-            resultView.markEnd();
         }
     }
 }
